main functionality ready

+++ server/gateway/src/resolvers/cart.resolver.ts
@@ -1,12 +1,26 @@
-import { Arg, Ctx, Mutation, Resolver } from "type-graphql";
+import { Arg, Ctx, Mutation, Query, Resolver } from "type-graphql";
 import axios from "axios";
-import { Context } from "../utils/types";
-import CartItem, { CartItemResponse } from "../types/cart-item.type";
-
 import { urls } from "../constants";
+import { Context } from "../utils/types";
+import { CartItemResponse } from "../types/cart-item.type";
+import Cart, { CartResponse } from "../types/cart.type";
+import { OrderResponse } from "../types/order.type";
 
-@Resolver(returns => CartItem)
-export default class CartItemResolver {
+
+@Resolver()
+export default class CartResolver {
+
+    @Query(returns => Cart)
+    async getCart(@Ctx() { req }: Context) {
+        if (!req.session.total) req.session.total = 0;
+        const response = await axios.get(urls.cart + 'cart-items?sid=' + req.sessionID)
+            .catch(error => { throw new Error(error.response.data.message) });
+        return {
+            sid: req.sessionID,
+            total: req.session.total,
+            items: response.data
+        }
+    }
 
     @Mutation(returns => CartItemResponse)
     async addItemToCart(
@@ -20,8 +34,10 @@ export default class CartItemResolver {
             let response = await axios.get(urls.catalog + 'variant/' + slug)
                 .catch(error => { throw new Error(error.response.data.message) })
             const variant = response.data;
-
-            if (variant.availability - qty < 0)
+            response = await axios.get(urls.cart + 'cart-items/getTaken/' + variant._id)
+                .catch(error => { throw new Error(error.response.data.message) });
+            const availability = variant.stock - variant.committed - response.data;
+            if (availability - qty < 0)
                 return {
                     error: {
                         field: 'qty',
@@ -38,10 +54,8 @@ export default class CartItemResolver {
                 }
             }).catch(error => { throw new Error(error.response.data.message) })
             const item = response.data;
-
-            req.session.total += item.total;
+            req.session.total += variant.price * qty;
             return { item };
-
         } catch (error) {
             return {
                 error: {
@@ -51,7 +65,6 @@ export default class CartItemResolver {
             }
         }
     }
-
 
     @Mutation(returns => CartItemResponse)
     async removeItemFromCart(
@@ -89,10 +102,8 @@ export default class CartItemResolver {
                 }
             }).catch(error => { throw new Error(error.response.data.message) })
             const item = response.data;
-
-            req.session.total -= item.total;
+            req.session.total -= variant.price * qty;
             return { item };
-
         } catch (error) {
             return {
                 error: {
@@ -103,13 +114,59 @@ export default class CartItemResolver {
         }
     }
 
-    @Mutation(returns => [CartItem])
+    @Mutation(returns => Boolean)
     async purgeItemsFromCart(
         @Ctx() { req }: Context
     ) {
-        const response = await axios.delete(urls.cart + 'cart/' + req.sessionID)
+        await axios.delete(urls.cart + 'cart/' + req.sessionID)
             .catch(error => { throw new Error(error.response.data.message) })
         req.session.total = 0;
-        return response.data;
+        return true;
+    }
+
+    // it is supposed to be order
+    @Mutation(returns => OrderResponse)
+    async saveOrder(@Ctx() { req }: Context) {
+        if (!req.session.total || req.session.total === 0)
+            return {
+                error: {
+                    field: 'total',
+                    message: 'Empty cart cannot be saved'
+                }
+            }
+        try {
+            // get cart items
+            let response = await axios.get(urls.cart + 'cart-items?sid=' + req.sessionID)
+                .catch(error => { throw new Error(error.response.data.message) });
+            const items = response.data;
+
+            // create order
+            response = await axios.post(urls.order, {
+                sid: req.sessionID,
+                total: req.session.total,
+                items
+            }).catch(error => { throw new Error(error.response.data.message) });
+
+            // update commited
+            for await (const item of items) {
+                axios.patch(urls.catalog + 'variant/' + item.product_variant + '?qty=' + item.quantity);
+            }
+
+            // delete cart-items
+            await axios.delete(urls.cart + 'cart/' + req.sessionID)
+                .catch(error => { throw new Error(error.response.data.message) })
+            req.session.destroy(err => {
+                if (err) throw new Error(err);
+            })
+            return { order: response.data }
+        } catch (error) {
+            return {
+                error: {
+                    field: '...',
+                    message: error.message
+                }
+            }
+        }
+
     }
 }
